+++ src/middleware/validation.js
@@ -1,7 +1,6 @@
 const validation = (schema, property = 'body') => {
   return (req, res, next) => {
     const { error, value } = schema.validate(req[property]);
-    console.log(req.body);
 
     const isValid = error == null;
 
@@ -12,7 +11,7 @@ const validation = (schema, property = 'body') => {
       const { details } = error;
       const message = details.map((i) => i.message).join(',');
 
-      console.log('error', message);
+      console.error('Validation Error -> ', message);
       res.status(422).json({ error: message });
     }
   };
